rugguard_env/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-rugguard-env"
version = "0.1.0"
description = "RugGuard — Crypto token scam detection environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "rugguard_env.server.app:main"

[tool.setuptools]
packages = ["rugguard_env", "rugguard_env.server"]
package-dir = { "rugguard_env" = ".", "rugguard_env.server" = "server" }
